spread the factory in files
