+++ package.json
@@ -38,7 +38,6 @@
     "eslint-loader": "^1.9.0",
     "eslint-plugin-html": "^3.2.1",
     "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-react": "^7.3.0",
     "express": "^4.15.4",
     "isomorphic-fetch": "^2.2.1",
@@ -62,6 +61,7 @@
     "webpack-dev-server": "^2.8.1"
   },
   "dependencies": {
+    "eslint-plugin-jsx-a11y": "5.1.1",
     "markdown-to-ast": "^4.0.0",
     "prop-types": "^15.5.10",
     "eslint-plugin-jsx-a11y": "^5.1.1",
